Renamed prefix convention into underscore convention.
Noticed that removing the L_ prefix would cause issues with PayPal since
they utilize it to not only flag list values, but also make the key
unique to same values without the L_ prefix. Thereby, the convention is
more related to the underscores in the key and we should not deal with
the L_ prefix at all. Instead we should treat it like any other key
component.