Fixed IndexError exception issues when given keys imposing list values.
The API response to GetExpressCheckoutDetails in PayPal can return keys
which are imposing sequential values, e.g PAYMENTREQUEST_n_SHIPTOSTREET
and PAYMENTREQUEST_n_SHIPTOSTREET2. This of course caused an IndexError
exception due to the index 2 being out of range in this case.

Such cases where, by design, intended to cause such an exception to be
raised. Since the assumption was that all keys (in the underscore
convention) with appended integers where intended to be lists in PayPal
responses. However, the GetExpressCheckoutDetails proved otherwise and
thus the new approach was implemented: Treating SHIPTOSTREET2 as a
dictionary key in case SHIPTOSTREET is not a list.