Fixed issue with out of range indexes in sequence impostors.
The previous solution was flawed since it did not check whether the
destination had the current key. Since we catched IndexError this did
not cause any issues in case the destination was a sequence. However, in
case it was a dictionary a KeyError exception would be raised - which we
did not handle.